move read/write tests to separate file
